delete duplicated Timer interfaces
