+++ relay/relay.go
@@ -46,14 +46,6 @@ const (
 	pollingDurationOnError   = time.Second * 5
 )
 
-type Timer interface {
-	After(d time.Duration) <-chan time.Time
-	NowUnix() int64
-	Start()
-	Close() error
-	IsInterfaceNil() bool
-}
-
 type defaultTimer struct {
 	ntpSyncTimer ntp.SyncTimer
 }
@@ -86,7 +78,7 @@ func (s *defaultTimer) IsInterfaceNil() bool {
 
 type Relay struct {
 	messenger relayp2p.NetMessenger
-	timer     Timer
+	timer     coreBridge.Timer
 	log       logger.Logger
 
 	ethBridge    bridge.Bridge
